# openapi.yaml
openapi: 3.0.1
info:
  title: Flippy
  description: '**Flippy**'
  contact:
    name: StackRoute
    url: http://stackroute.in
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: http://localhost:7000
  description: Local Server
- url: http://dev-flippy.com
  description: Dev
- url: http://flippy.com
  description: Prod
...
security:
  - Authorization_Header: []
paths:
  /api/orders:
    post:
      tags:
      - Flippy
      summary: 'Place order (Returns: CustomerOrder)'
      operationId: createOrder
      requestBody:
        description: CustomerOrder object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerOrder'
        required: true
      responses:
        200:
          description: Successful operation
          content: {}
        405:
          description: Invalid input
          content: {}
        default:
          description: Unexpected error
          content: {}
    get:
      tags:
      - Flippy
      summary: 'Get all Customer Orders (paginated) (Returns: CustomerOrdersList)'
      parameters:
      - name: customerId
        in: query
        description: CustomerId that need to be considered for filter
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: per_page
        in: query
        description: Number of items per page
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      - name: page
        in: query
        description: The index of page to be considered for filter
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerOrder'
          headers:
            X-Total:
              description: The total number of items 
              schema:
                type: integer
            X-Total-Pages:
              description: The total number of pages  
              schema:
                type: integer
            X-Per-Page:
              description: The number of items per page
              schema:
                type: integer
            X-Page:
              description: The index of the current page (starting at 1)
              schema:
                type: integer
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Customer not found
          content: {}
...
  /api/shop/orders:
    get:
      tags:
      - Flippy Shop Partner
      summary: 'Get all Shop Orders (paginated) (Returns: ShopOrdersList)'
      parameters:
      - name: shopId
        in: query
        description: ShopId that need to be considered for filter
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: per_page
        in: query
        description: Number of items per page
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      - name: page
        in: query
        description: The index of page to be considered for filter
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopOrder'
          headers:
            X-Total:
              description: The total number of items 
              schema:
                type: integer
            X-Total-Pages:
              description: The total number of pages  
              schema:
                type: integer
            X-Per-Page:
              description: The number of items per page
              schema:
                type: integer
            X-Page:
              description: The index of the current page (starting at 1)
              schema:
                type: integer
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Customer not found
          content: {}
...
components:
  schemas:
    CustomerOrder: {}
    ShopOrder: {}
...
  securitySchemes:
    apiKey:
      type: apiKey
      name: api-key
      in: header